backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
insightface==0.7.3
onnxruntime==1.17.0
ffmpeg-python==0.2.0
numpy==1.26.3
pillow==10.2.0
python-multipart==0.0.6
scikit-learn==1.4.0
aiofiles==23.2.1
